\name{Sling-package}
\alias{Sling-package}
\alias{Sling}
\docType{package}
\title{
\packageTitle{Sling}
}
\description{
\packageDescription{Sling}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{Sling}
\packageIndices{Sling}
}
\author{
\packageAuthor{Sling}
Maintainer: \packageMaintainer{Sling}
}

\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
